+++ js/bingo/mv/ajax-vsdoc.js
@@ -103,7 +103,9 @@
             //是否包函url query部分作为key 缓存数据, 默认true
             cacheQurey: true,
             //hold server数据, function(response, isSuccess, xhr){return return [response, isSuccess, xhr];}
-            holdServer: null
+            holdServer: null,
+            //处理参数, function(){ return this.param()}
+            holdParams: null
         });
 
         this.Define({